fix: remove deprecated iAd Attribution API
